Make tenantId optional, add password field

+++ repository/queries/account/accountDelete.sql
@@ -1,14 +1,7 @@
 
 WITH deleted AS
 	(
-		DELETE FROM
-			account.account
-			
-		WHERE
-					tenant_id=$1
-			AND id=$2
-		
-		RETURNING *
+		DELETE FROM account.account WHERE ($1::bigint IS NULL OR tenant_id=$1::bigint) AND id=$2 RETURNING *
 	)
 	
 SELECT
